pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chitrakatha"
version = "0.1.0"
description = "Serverless MLOps platform for Indian Comic History LLM"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34.0",
    "sagemaker>=2.200.0,<3.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.2.0",
    "transformers>=4.38.0",
    "peft>=0.9.0",
    "trl>=0.8.0",
    "datasets>=2.17.0",
    "bitsandbytes>=0.43.0",
    "yt-dlp>=2024.3.0",
    "openpyxl>=3.1.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "numpy>=1.26.0",
    "pytz",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "moto[s3,bedrock,secretsmanager]>=5.0.0"
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
